refactor: add corrections
